Code Files move to parent foler. Code updated to avoid Analyzer warnings and error.
