--- com.iPAHeartBeat.Core.Extensions/Scripts/iPAHeartBeat.Core.Extensions.asmdef
@@ -1,16 +1,0 @@
-{
-    "name": "iPAHeartBeat.Core.Extensions",
-    "rootNamespace": "iPAHeartBeat.Core.Extensions",
-    "references": [],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": true,
-    "precompiledReferences": [
-        "Newtonsoft.Json.dll"
-    ],
-    "autoReferenced": true,
-    "defineConstraints": [],
-    "versionDefines": [],
-    "noEngineReferences": true
-}
