+++ iPAHeartBeat.Core/Extensions/JsonExtensions.cs
@@ -1,0 +1,86 @@
+// --------------------------------------------------------------------------------
+// <copyright file="JsonExtensions.cs" company="N/A">
+// Copyright (c) N/A. All rights reserved.
+// </copyright>
+// --------------------------------------------------------------------------------
+
+using System;
+using System.Collections;
+
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
+
+namespace iPAHeartBeat.Core.Extensions;
+
+/// <summary>
+/// Extension class to work with json in Unity. this will use Unity Json Utility and MiniJson from Calvin Rain.
+/// </summary>
+public static class JsonExtensions {
+	/// <summary>
+	/// Gets readonly Newton soft Json Serializer settings. To Update setting please use <see cref="SetSerializerSettings(JsonSerializerSettings)"/>.
+	/// </summary>
+	public static JsonSerializerSettings Settings { get; private set; } = new JsonSerializerSettings();
+
+	/// <summary>
+	/// Gets readonly Json Serializer formatting option.
+	/// </summary>
+	public static Formatting FormattingOption { get; private set; } = Formatting.None;
+
+	/// <summary>
+	/// Will update Newton Soft Json Serializer Setting to serialize and deserialize JSon data.
+	/// <br/> NOTE: Please use same setting or default setting in whole game to avoid any kind of mismatch.
+	/// </summary>
+	/// <param name="settings">The Serialization setting option to set for NewtonSoft Json.</param>
+	public static void SetSerializerSettings(JsonSerializerSettings settings) {
+		if (settings.IsNull()) {
+			throw new ArgumentNullException(nameof(settings));
+		}
+
+		Settings = settings;
+	}
+
+	/// <summary>
+	/// Will Serialized object as JSon string using Unity Json Utility with pretty format.
+	/// </summary>
+	/// <param name="obj">object to serialized.</param>
+	/// <returns>serialized JSon string.</returns>
+	public static string ToJson(this object obj)
+		=> obj.ToJson(FormattingOption);
+
+	/// <summary>
+	/// Will Serialized object as JSon string using Unity Json Utility with pretty format.
+	/// </summary>
+	/// <param name="obj">object to serialized.</param>
+	/// <param name="formatting">Serializer formatting option like convert Object to JSON string with indentation or without indent like minified JSON string.</param>
+	/// <returns>serialized JSon string.</returns>
+	public static string ToJson(this object obj, Formatting formatting)
+		=> JsonConvert.SerializeObject(obj, formatting);
+
+	/// <summary>
+	/// will Deserialize json string to Dictionary.
+	/// </summary>
+	/// <typeparam name="T">Managed Type to which Json data needs to convert.</typeparam>
+	/// <param name="jsonData">string to deserialize.</param>
+	/// <returns>Dict data from serialized json string.</returns>
+	public static T FromJson<T>(this string jsonData)
+		=> JsonConvert.DeserializeObject<T>(jsonData);
+
+	/// <summary>
+	/// The helper method to Convert Object from one type to another type using json serialization.
+	/// </summary>
+	/// <typeparam name="T">Managed Type to which Input data needs to convert.</typeparam>
+	/// <param name="obj">The object which needs to Convert.</param>
+	/// <returns>instance of <typeparamref name="T"/> with data in it.</returns>
+	/// <exception cref="InvalidCastException">This method will throw exception if input object is primitive data type like <see cref="string"/>, <see cref="int"/>. etc...</exception>
+	public static T ConvertTo<T>(this object obj) {
+		var rs = obj switch {
+			JObject njObj => njObj.ToString(),
+			IDictionary dict => dict.ToJson(),
+			_ => obj.IsPrimitiveType()
+				? throw new InvalidCastException("Primitive data type can't converted to Json.")
+				: obj.ToJson(),
+		};
+
+		return JsonConvert.DeserializeObject<T>(rs);
+	}
+}
